pyproject.toml
[project]
name = "minha-receita"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "isort>=6.0.1",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.3",
    "ruff>=0.11.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
linter = [
    "isort>=6.0.1",
    "ruff>=0.11.2",
]
